refactored all components, services, and models
