+++ src/app/app.component.html
@@ -1,1 +1,1 @@
-<app-view></app-view>
+<app-view-associate></app-view-associate>
