+++ src/app/components/view/view-associate.component.css
