+++ src/app/components/view/view-associate.component.html
