+++ src/app/models/associate-model/associate.model.spec.ts
@@ -1,0 +1,2 @@
+import { Associate } from './associate.model';
+
