--- src/app/models/associate.model.spec.ts
@@ -1,7 +1,0 @@
-import { Associate } from './associate.model';
-
-describe('Associate', () => {
-  it('should create an instance', () => {
-    expect(new Associate()).toBeTruthy();
-  });
-});
